refactor: 🚚 ゲームに関するロジックを役割ごとにファイルに切り出しました
